Move the Customized User model into MyAccount app and did the relevant other changes.
