+++ src/profile/__init__.py
@@ -1,2 +1,1 @@
-
 default_app_config = 'profile.apps.ProfileConfig'
