+++ src/profile/admin.py
@@ -1,19 +1,8 @@
 # -*- coding: utf-8 -*-
 
 from django.contrib import admin
-from django.contrib.auth.admin import UserAdmin
-from django.utils.translation import ugettext_lazy as _
 
-from profile.models import (
-    Profile,
-    User
-)
-
-from profile.forms import (
-    UserChangeForm,
-    UserCreationForm,
-    AdminPasswordChangeForm
-)
+from .models import Profile
 
 
 @admin.register(Profile)
@@ -25,29 +14,4 @@ class ProfileAdmin(admin.ModelAdmin):
         if obj.description:
             short_des = obj.description[:50]
         return short_des
-
-
-@admin.register(User)
-class UserAdmin(UserAdmin):
-    fieldsets = (
-        (None, {'fields': ('email', 'password')}),
-        (_('Personal info'), {'fields': ('full_name', 'mobile')}),
-        (_('Permissions'), {'fields': ('is_active', 'is_staff', 'is_superuser',
-                                       'groups', 'user_permissions')}),
-        (_('Important dates'), {'fields': ('last_login', 'date_joined')}),
-    )
-    add_fieldsets = (
-        (None, {
-            'classes': ('wide',),
-            'fields': ('email', 'password1', 'password2'),
-        }),
-    )
-    form = UserChangeForm
-    add_form = UserCreationForm
-    change_password_form = AdminPasswordChangeForm
-    list_display = ('email', 'full_name', 'is_staff')
-    list_filter = ('is_staff', 'is_superuser', 'is_active', 'groups')
-    search_fields = ('username', 'full_name', 'email')
-    ordering = ('email',)
-    filter_horizontal = ('groups', 'user_permissions',)
