+++ src/profile/apps.py
@@ -4,7 +4,6 @@ from django.apps import AppConfig
 from django.db.models.signals import (
     post_save,
 )
-from django.conf import settings
 from django.utils.translation import gettext_lazy as _
 
 
@@ -13,15 +12,18 @@ class ProfileConfig(AppConfig):
     verbose_name = _("Profile")
 
     def ready(self):
-        from profile.models import (
+        from django.contrib.auth import get_user_model
+        from .models import (
             post_create_profile_handler,
             post_update_profile_handler
         )
 
+        User = get_user_model()
+
         post_save.connect(
-            post_create_profile_handler, sender=settings.AUTH_USER_MODEL,
+            post_create_profile_handler, sender=User,
             dispatch_uid='profile-post-save-user')
 
         post_save.connect(
-            post_update_profile_handler, sender=settings.AUTH_USER_MODEL,
+            post_update_profile_handler, sender=User,
             dispatch_uid='profile-post-update-user')
